pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "fastapi-utils>=0.8.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.7",
    "langgraph>=1.0.6",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "typing-inspect>=0.9.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = ["pytest>=9.0.2"]


[tool.setuptools]
package-dir = { "" = "src" }
packages = ["app", "auth", "common", "configs"]


[tool.uv]
index-url = "https://pypi.org/simple"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]


[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app", "auth", "common", "configs"]
